refactor(post): 게시글 작성시 요청 DTO 네이밍 변경
