+++ src/main/java/org/sopt/domain/user/dto/request/CreateUserRequest.java
@@ -1,4 +1,4 @@
-package org.sopt.domain.user.dto;
+package org.sopt.domain.user.dto.request;
 
 public record CreateUserRequest(String name) {
 }
